requirements.txt
tensorflow==2.17.1
streamlit
pillow
numpy
pandas